refactor: Move utils to util.js
